added foam support
